325: converted withNbdr as query param and shifted blockspan operation into dedicated service
