+++ backend/src/oe/api/op-energy.service.ts
@@ -3,13 +3,11 @@ import crypto from "crypto-js";
 import bitcoinApi from '../../api/bitcoin/bitcoin-api-factory';
 import { IEsploraApi } from '../../api/bitcoin/esplora-api.interface';
 import opBlockHeaderService from '../service/op-block-header.service';
-import opStatisticsService from './op-statistics.service';
 import opBlockHeaderRepository from '../repositories/OpBlockHeaderRepository';
 import config from '../../config';
 import logger from '../../logger';
 
 
-import { NbdrStatisticsError } from "./interfaces/op-statistics.interface";
 import {
   BlockHeader,
   SlowFastGuessValue,
@@ -26,11 +24,7 @@ import {
   TimeStrikesHistory,
   SlowFastResult,
   BlockHash,
-  BlockSpan,
   RegisterResponse,
-  ConfirmedBlockHeight,
-  BlockSpanDetails,
-  BlockSpanDetailsWithNbdr
 } from './interfaces/op-energy.interface';
 
 // magic values, that should be contained by account token hash in order to pass a quick check
@@ -521,140 +515,6 @@ export class OpEnergyApiService {
   public async $getBlockByHash( hash: BlockHash): Promise<IEsploraApi.Block> {
     // using our own block cache goes here
     return await bitcoinApi.$getBlock(hash.value);
-  }
-
-  public $getBlockSpanList(UUID: string, endBlockHeight: number, span: number, blockSpanCount: number): BlockSpan[] {
-    try {
-      const blockSpanList = [] as BlockSpan[];
-      const numberOfSpan = blockSpanCount === -1 ? Number.MAX_VALUE : blockSpanCount;
-      for (let i = endBlockHeight; (i - span) > 6 && blockSpanList.length < numberOfSpan; i -= span) {
-        const blockSpan = {
-          startBlockHeight: i - span,
-          endBlockHeight: i
-        };
-        blockSpanList.push(blockSpan);
-      }
-      return blockSpanList;
-    } catch (e) {
-      throw new Error(`${UUID} OpEnergyApiService.$getBlockSpanList: error while generating block span list: ${e instanceof Error ? e.message : e}`);
-    }
-  }
-
-  private async $getConfirmedBlockSpanList(UUID: string, blockSpanList: BlockSpan[]): Promise<ConfirmedBlockHeight[]> {
-    try {
-      const confirmedBlocks = [] as ConfirmedBlockHeight[];
-      const currentTip = await bitcoinApi.$getBlockHeightTip();
-
-      blockSpanList.forEach((blockNumber) =>
-        confirmedBlocks.push(
-          opBlockHeaderService.verifyConfirmedBlockHeight(
-            blockNumber.startBlockHeight,
-            {
-              value: currentTip,
-            }
-          ),
-          opBlockHeaderService.verifyConfirmedBlockHeight(
-            blockNumber.endBlockHeight,
-            {
-              value: currentTip,
-            }
-          )
-        )
-      );
-      return confirmedBlocks;
-    } catch (e) {
-      throw new Error(`${UUID} OpEnergyApiService.$getConfirmedBlockSpanList: error while filtering confirmed block span list: ${e instanceof Error ? e.message : e}`);
-    }
-  }
-
-  public async $generateBlockSpanDetailedList(UUID: string, endBlockHeight: number, span: number, blockSpanCount: number): Promise<BlockSpanDetails[]> {
-    try {
-      const blockSpanList = this.$getBlockSpanList(
-        UUID,
-        endBlockHeight,
-        span,
-        blockSpanCount
-      );
-      const confirmedBlockSpanList = await this.$getConfirmedBlockSpanList(UUID, blockSpanList);
-      if (!confirmedBlockSpanList.length) {
-        return [];
-      }
-
-      const blockHeadersList = await opBlockHeaderService.$getBlockHeadersByHeights(
-        UUID,
-        Array.from(new Set(confirmedBlockSpanList))
-      );
-
-      let index = 0;
-      const result: BlockSpanDetails[] = [];
-      while (index < blockHeadersList.length - 1) {
-        result.push({
-          startBlock: {
-            height: blockHeadersList[index].height,
-            hash: blockHeadersList[index].current_block_hash,
-            mediantime: blockHeadersList[index].mediantime,
-            timestamp: blockHeadersList[index].timestamp,
-          },
-          endBlock: {
-            height: blockHeadersList[++index].height,
-            hash: blockHeadersList[index].current_block_hash,
-            mediantime: blockHeadersList[index].mediantime,
-            timestamp: blockHeadersList[index].timestamp,
-          },
-        });
-      }
-      return result;
-    } catch (e) {
-      throw new Error(`${UUID} OpEnergyApiService.$generateBlockSpanDetailedList: error while generating detailed block span list: ${e instanceof Error ? e.message : e}`);
-    }
-  }
-
-  public async $getBlockSpanDetailedList(
-    UUID: string,
-    endBlockHeight: number,
-    span: number,
-    blockSpanCount: number,
-    withNbdrStatistics: boolean
-  ): Promise<BlockSpanDetails[] | BlockSpanDetailsWithNbdr[] | NbdrStatisticsError> {
-    try {
-      if (blockSpanCount !== -1 && blockSpanCount <= 0) {
-        throw new Error('Invalid block span count');
-      }
-
-      const blockSpanDetailedList = await this.$generateBlockSpanDetailedList(
-        UUID,
-        endBlockHeight,
-        span,
-        withNbdrStatistics && blockSpanCount !== -1 ? blockSpanCount + 100 : blockSpanCount
-      );
-      if (!withNbdrStatistics) {
-        return blockSpanDetailedList;
-      }
-
-      for (let index = 100; index < blockSpanDetailedList.length; index++) {
-        const blockSpanListForStat = blockSpanDetailedList.slice(index - 100, index);
-        const statistics = opStatisticsService.calculateStatisticsWithBlockSpan(blockSpanListForStat, span);
-        if ('error' in statistics) {
-          return statistics;
-        }
-        blockSpanDetailedList[index]['nbdr'] = {
-          avg: statistics.nbdr.avg,
-          stddev: statistics.nbdr.stddev
-        };
-      }
-
-      if (blockSpanCount === -1) {
-        return blockSpanDetailedList;
-      }
-
-      if (blockSpanDetailedList.length > 100) {
-        return blockSpanDetailedList.slice(100);
-      }
-
-      return blockSpanDetailedList.slice(blockSpanDetailedList.length - blockSpanCount);
-    } catch (e) {
-      throw new Error(`${UUID} OpEnergyApiService.$getBlockSpanDetailedList: error while fetching detailed block span list: ${e instanceof Error ? e.message : e}`);
-    }
   }
 
   // this procedure generates new random account secret and it's token, which is really a hash
