+++ backend/src/oe/util/helper.ts
@@ -7,10 +7,6 @@ export const isValidNaturalNumber = (numberString: string): boolean => {
   return parseInt(numberString) >= 0;
 };
 
-export const isBoolean = (expression: string): boolean => {
-  return ['true', 'false'].includes(expression);
-};
-
 export const toBoolean = (expression: string): boolean => {
   return expression.toLowerCase() === 'true';
 };
